wedding_vendors_web/requirements.txt
beautifulsoup4==4.9.1
Django==5.0.6
djangorestframework==3.15.2
Requests==2.32.3
selenium==4.26.1
